+++ projects/ccd-case-ui-toolkit/src/lib/shared/components/case-viewer/printer/case-printer.component.spec.ts
@@ -7,6 +7,7 @@ import { AbstractAppConfig } from '../../../../app.config';
 import { CasePrintDocument, CaseView } from '../../../domain';
 import { AlertService } from '../../../services';
 import { attr, text } from '../../../test/helpers';
+import { MockRpxTranslatePipe } from '../../../test/mock-rpx-translate.pipe';
 import { CaseNotifier, CasesService } from '../../case-editor';
 import { PaletteUtilsModule } from '../../palette';
 import { CasePrinterComponent } from './case-printer.component';
@@ -89,7 +90,8 @@ describe('CasePrinterComponent', () => {
           PrintUrlPipe,
 
           // Mock
-          CaseHeaderComponent
+          CaseHeaderComponent,
+          MockRpxTranslatePipe
         ],
         providers: [
           { provide: CaseNotifier, useValue: caseService },