+++ projects/ccd-case-ui-toolkit/src/lib/shared/components/palette/base-field/field-read.component.spec.ts
@@ -8,6 +8,7 @@ import { of } from 'rxjs';
 import { CaseEditDataService } from '../../../commons/case-edit-data';
 import { CaseEventData } from '../../../domain/case-event-data.model';
 import { CaseField } from '../../../domain/definition';
+import { MockRpxTranslatePipe } from '../../../test/mock-rpx-translate.pipe';
 import { Draft } from '../../../domain/draft.model';
 import { CaseFieldService } from '../../../services/case-fields/case-field.service';
 import { FieldTypeSanitiser } from '../../../services/form/field-type-sanitiser';
@@ -19,6 +20,7 @@ import { PageValidationService } from '../../case-editor/services/page-validatio
 import { PaletteService } from '../palette.service';
 import { FieldReadComponent } from './field-read.component';
 import { PaletteContext } from './palette-context.enum';
+
 
 import createSpyObj = jasmine.createSpyObj;
 
@@ -140,6 +142,7 @@ describe('FieldReadComponent', () => {
           // Mock
           FieldTestComponent,
           FieldReadLabelComponent,
+          MockRpxTranslatePipe,
         ],
         providers: [
           { provide: PaletteService, useValue: paletteService },