+++ projects/ccd-case-ui-toolkit/src/lib/shared/components/palette/case-flag/read-case-flag-field.component.spec.ts
@@ -12,6 +12,7 @@ import { FlagDetail, FlagDetailDisplay, FlagsWithFormGroupPath } from './domain'
 import { CaseFlagStatus, CaseFlagSummaryListDisplayMode } from './enums';
 import { ReadCaseFlagFieldComponent } from './read-case-flag-field.component';
 import { WriteCaseFlagFieldComponent } from './write-case-flag-field.component';
+import { MockRpxTranslatePipe } from '../../../test/mock-rpx-translate.pipe';
 
 describe('ReadCaseFlagFieldComponent', () => {
   let component: ReadCaseFlagFieldComponent;
@@ -373,9 +374,9 @@ describe('ReadCaseFlagFieldComponent', () => {
     caseFlagStateServiceSpy.pageLocation = '../createCaseFlag/createCaseFlagCaseFlagFormPage';
 
     TestBed.configureTestingModule({
-      schemas: [ CUSTOM_ELEMENTS_SCHEMA ],
       imports: [ RouterTestingModule ],
-      declarations: [ ReadCaseFlagFieldComponent ],
+      schemas: [ CUSTOM_ELEMENTS_SCHEMA ],
+      declarations: [ ReadCaseFlagFieldComponent, MockRpxTranslatePipe ],
       providers: [
         { provide: ActivatedRoute, useValue: mockRoute },
         { provide: CaseFlagStateService, useValue: caseFlagStateServiceSpy },