+++ projects/ccd-case-ui-toolkit/src/lib/shared/components/palette/collection/write-collection-field.component.spec.ts
@@ -11,6 +11,7 @@ import { CaseField, FieldType } from '../../../domain/definition';
 import { createAProfile } from '../../../domain/profile/profile.test.fixture';
 import { ProfileNotifier } from '../../../services';
 import { FormValidatorsService } from '../../../services/form';
+import { MockRpxTranslatePipe } from '../../../test/mock-rpx-translate.pipe';
 import { RemoveDialogComponent } from '../../dialogs/remove-dialog';
 import { PaletteUtilsModule } from '../utils';
 import { CollectionCreateCheckerService } from './collection-create-checker.service';
@@ -114,6 +115,7 @@ describe('WriteCollectionFieldComponent', () => {
         ],
         declarations: [
           WriteCollectionFieldComponent,
+          MockRpxTranslatePipe,
           fieldWriteComponent,
           fieldReadComponent
         ],
@@ -362,6 +364,7 @@ describe('WriteCollectionFieldComponent CRUD impact', () => {
         ],
         declarations: [
           WriteCollectionFieldComponent,
+          MockRpxTranslatePipe,
           fieldWriteComponent,
           fieldReadComponent
         ],
@@ -477,6 +480,7 @@ describe('WriteCollectionFieldComponent CRUD impact - Update False', () => {
         ],
         declarations: [
           WriteCollectionFieldComponent,
+          MockRpxTranslatePipe,
           fieldWriteComponent,
           fieldReadComponent
         ],
@@ -565,6 +569,7 @@ describe('WriteCollectionFieldComponent remove component from collection', () =>
         ],
         declarations: [
           WriteCollectionFieldComponent,
+          MockRpxTranslatePipe,
           fieldWriteComponent,
           fieldReadComponent
         ],