+++ projects/ccd-case-ui-toolkit/src/lib/shared/components/palette/yes-no/read-yes-no-field.component.spec.ts
@@ -3,9 +3,10 @@ import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { FormGroup } from '@angular/forms';
 import { CaseField } from '../../../domain/definition/case-field.model';
 import { FieldType } from '../../../domain/definition/field-type.model';
+import createSpyObj = jasmine.createSpyObj;
+import { MockRpxTranslatePipe } from '../../../test/mock-rpx-translate.pipe';
 import { ReadYesNoFieldComponent } from './read-yes-no-field.component';
 import { YesNoService } from './yes-no.service';
-import createSpyObj = jasmine.createSpyObj;
 
 describe('ReadYesNoFieldComponent', () => {
 
@@ -40,7 +41,8 @@ describe('ReadYesNoFieldComponent', () => {
         .configureTestingModule({
           imports: [],
           declarations: [
-            ReadYesNoFieldComponent
+            ReadYesNoFieldComponent,
+            MockRpxTranslatePipe
           ],
           providers: [
             {provide: YesNoService, useValue: yesNoService}
@@ -86,7 +88,8 @@ describe('ReadYesNoFieldComponent', () => {
         .configureTestingModule({
           imports: [],
           declarations: [
-            ReadYesNoFieldComponent
+            ReadYesNoFieldComponent,
+            MockRpxTranslatePipe
           ],
           providers: [
             {provide: YesNoService, useValue: yesNoService}