+++ projects/ccd-case-ui-toolkit/src/lib/shared/components/search-result/search-result.component.spec.ts
@@ -18,6 +18,7 @@ import {
 } from '../../domain';
 import { CaseReferencePipe, SortSearchResultPipe } from '../../pipes';
 import { ActivityService, BrowserService, FieldsUtils, SearchResultViewItemComparatorFactory, SessionStorageService } from '../../services';
+import { MockRpxTranslatePipe } from '../../test/mock-rpx-translate.pipe';
 import { SearchResultComponent } from './search-result.component';
 import createSpyObj = jasmine.createSpyObj;
 
@@ -194,6 +195,8 @@ describe('SearchResultComponent', () => {
             SortSearchResultPipe,
             CaseReferencePipe,
             // Mocks
+            PaginatePipe,
+            MockRpxTranslatePipe,
             caseActivityComponent,
             PaginatePipe
           ],
@@ -821,6 +824,8 @@ describe('SearchResultComponent', () => {
             SortSearchResultPipe,
             CaseReferencePipe,
             // Mocks
+            PaginatePipe,
+            MockRpxTranslatePipe,
             caseActivityComponent,
             PaginatePipe
           ],