+++ projects/ccd-case-ui-toolkit/src/lib/shared/pipes/index.ts
@@ -2,5 +2,7 @@ export * from './case-reference';
 export * from './case-title';
 export * from './complex';
 export * from './generic';
+export * from './link-cases-reason-code';
+export * from './link-cases-from-reason-code';
 export * from './pipes.module';
 export * from './search-result/sorting/sort-search-result.pipe';