+++ projects/ccd-case-ui-toolkit/src/lib/shared/pipes/pipes.module.ts
@@ -4,6 +4,8 @@ import { CaseReferencePipe } from './case-reference';
 import { CcdCaseTitlePipe } from './case-title';
 import { CcdCollectionTableCaseFieldsFilterPipe, CcdCYAPageLabelFilterPipe, CcdPageFieldsPipe, CcdTabFieldsPipe, ReadFieldsFilterPipe } from './complex';
 import { EnumDisplayDescriptionPipe } from './generic/enum-display-description/enum-display-description.pipe';
+import { LinkCasesFromReasonValuePipe } from './link-cases-from-reason-code/ccd-link-cases-from-reason-code.pipe';
+import { LinkCasesReasonValuePipe } from './link-cases-reason-code/ccd-link-cases-reason-code.pipe';
 import { SortSearchResultPipe } from './search-result/sorting/sort-search-result.pipe';
 
 const pipeDeclarations = [
@@ -15,6 +17,8 @@ const pipeDeclarations = [
   ReadFieldsFilterPipe,
   CcdTabFieldsPipe,
   CcdPageFieldsPipe,
+  LinkCasesReasonValuePipe,
+  LinkCasesFromReasonValuePipe,
   EnumDisplayDescriptionPipe
 ];