+++ RELEASE-NOTES.md
@@ -22,6 +22,27 @@
 ### Version 6.10.7-case-flags-confirm-flag-status-step
 **EUI-7350** Add "confirm flag status" step of Create Case Flag journey
 
+### Version 6.13.10-rc1
+**EUI-5298** Release 6.13.10 including 6.13.0, linked cases (EUI-7676) and case file view fix (EUI-7417)
+
+### Version 6.13.0-rc11
+**EUI-7796** Re-tag for Release 6.13.0
+
+### Version 6.13.0-rc7
+**EUI-7796** Fix flag types retrieval to use HMCTSServiceId from supplementary data in the first instance (if available); this caters for case types that are associated with multiple HMCTS service codes
+
+### Version 6.10.10-linked-cases-other-description
+**EUI-7246** Linked cases other description
+
+### Version 6.10.10-linked-cases-and-case-file-view
+**EUI-7676** Configure linked cases
+**EUI-7417** Set boolean flag allowing Case File View document move functionality in readonly mode to `false`
+
+### Version 5.0.50-case-flags-multiple-service-codes-for-jurisdiction
+**EUI-7529** Fix flag types retrieval to use HMCTS service code associated with case type ID, falling back on jurisdiction only if the case type ID is not recognised, since there could be multiple service codes per jurisdiction
+
+**EUI-3729** Support combinations of AND and OR boolean logic
+
 ### Version 6.10.7-rollback-case-link
 **EUI-7675** Rollback case link
 
@@ -31,14 +52,14 @@
 ### Version 5.0.50-case-flags-multiple-case-flags-tabs-support
 **EUI-7688** Support multiple Case Flags tabs
 
+### Version 5.0.50-case-file-view-restrict-document-move
+**EUI-7417** Set boolean flag allowing Case File View document move functionality in readonly mode to `false`
+
 ### Version 5.0.50-sscs-sit-release
 **EUI-7564** SSCS SIT release
 
 ### Version 5.0.46-case-flags-not-approved-status
 **EUI-7347/EUI-7570** Support display of new "Not approved" status for Case Flags
-
-### Version 5.0.46-fix-case-link
-**EUI-7521** Fix case link hyperlink on applications tab
 
 ### Version 5.0.46-case-file-view-media-viewer-scrolling-behaviour
 **EUI-7496** Fix incorrect Media Viewer vertical scrolling behaviour beyond Case File View container bottom boundary
@@ -46,7 +67,7 @@
 ### Version 5.0.46-case-file-view-styling-fixes
 **EUI-7481** Case file view styling issues
 
-### Version 5.0.45-fix-case-acccess-table
+### Version 5.0.45-fix-case-access-table
 **EUI-7454** Fixed table alignment on case view
 
 ### Version 5.0.44-case-flags-integration-fixes